Add files via upload
Aim: To find the chromatic number for a graph using backtracking algorithm. The number of vertices in the graph and the adjacency matrix is provided as input by user.
The chromatic number of a graph is the smallest number of colors needed to color the vertices of so that no two adjacent vertices share the same color.